requirements.txt
python-telegram-bot[job-queue]>=20.7

sqlalchemy>=2.0

pytz>=2023.3
setuptools